Serveo

+++ server/msh/controller/user.py
@@ -27,9 +27,25 @@ class User(BaseHandler):
                             response['users'] = DbManager.select_tb_user(self.session.get('user'))
                         response['output'] = 'OK'
                     if tipo_operazione == "update":
-                        to_update = DbManager.select_tb_user(username)[0]
-                        if to_update['role'] != role:
-                            if self.session.get('role') == 'ADMIN':
+                        to_update = DbManager.select_tb_user(username)
+                        if len(to_update) == 1:
+                            to_update = to_update[0]
+                            if to_update['role'] != role:
+                                if self.session.get('role') == 'ADMIN':
+                                    if to_update['password'] != password:
+                                        if self.session.get('user') == username:
+                                            DbManager.update_tb_user(username, password, role)
+                                            update_user(username, password)
+                                            response['output'] = 'OK'
+                                        else:
+                                            response['output'] = 'Solo l\'utente propietario può modificare la sua password'
+                                    else:
+                                        DbManager.update_tb_user(username, password, role)
+                                        update_user(username, password)
+                                        response['output'] = 'OK'
+                                else:
+                                    response['output'] = 'Solo gli ADMIN possono modificare i ruoli'
+                            else:
                                 if to_update['password'] != password:
                                     if self.session.get('user') == username:
                                         DbManager.update_tb_user(username, password, role)
@@ -37,41 +53,33 @@ class User(BaseHandler):
                                         response['output'] = 'OK'
                                     else:
                                         response['output'] = 'Solo l\'utente propietario può modificare la sua password'
-                                else:
-                                    DbManager.update_tb_user(username, password, role)
-                                    update_user(username, password)
-                                    response['output'] = 'OK'
-                            else:
-                                response['output'] = 'Solo gli ADMIN possono modificare i ruoli'
                         else:
-                            if to_update['password'] != password:
-                                if self.session.get('user') == username:
-                                    DbManager.update_tb_user(username, password, role)
-                                    update_user(username, password)
-                                    response['output'] = 'OK'
-                                else:
-                                    response['output'] = 'Solo l\'utente propietario può modificare la sua password'
+                            response['output'] = 'Non esiste nessun utente con questo username'
                     if tipo_operazione == "delete":
                         users = DbManager.select_tb_user()
-                        to_delete = DbManager.select_tb_user(username)[0]
-                        admin = 0
-                        for user in users:
-                            if user['role'] == 'ADMIN':
-                                admin = admin + 1
-                        if to_delete['role'] == 'ADMIN' and admin == 1:
-                            response['output'] = 'Deve essere sempre presente almeno un utente ADMIN'
+                        to_delete = DbManager.select_tb_user(username)
+                        if len(to_delete) == 1:
+                            to_delete = to_delete[0]
+                            admin = 0
+                            for user in users:
+                                if user['role'] == 'ADMIN':
+                                    admin = admin + 1
+                            if to_delete['role'] == 'ADMIN' and admin == 1:
+                                response['output'] = 'Deve essere sempre presente almeno un utente ADMIN'
+                            else:
+                                DbManager.delete_tb_user(username)
+                                delete_user(username)
+                                response['output'] = 'OK'
+                        if tipo_operazione == "add":
+                            users = DbManager.select_tb_user(username)
+                            if len(users) == 0:
+                                DbManager.insert_tb_user(username, password, role)
+                                add_user(username, password)
+                                response['output'] = 'OK'
+                            else:
+                                response['output'] = 'Username già utilizzato'
                         else:
-                            DbManager.delete_tb_user(username)
-                            delete_user(username)
-                            response['output'] = 'OK'
-                    if tipo_operazione == "add":
-                        users = DbManager.select_tb_user(username)
-                        if len(users) == 0:
-                            DbManager.insert_tb_user(username, password, role)
-                            add_user(username, password)
-                            response['output'] = 'OK'
-                        else:
-                            response['output'] = 'Username già utilizzato'
+                            response['output'] = 'Non esiste nessun utente con questo username'
                     DbManager.close_db()
                 else:
                     response['output'] = 'La funzione richiesta può essere eseguita solo da un ADMIN'
